+++ changelog.md
@@ -1,9 +1,15 @@
 # Changelog
 
-## Version 0.0 (*July 2022*) 
+## Version 0.0 (_July 2022_)
 
 * Rebranded from POTD to the Community Server Bot
 * Created `changelog.md`!
 * Shortened problem IDs to 6 digits from 9.
 * Updated `README.md` to reflect changes
 * Scoring formula entirely redone (decreases from 1000 points to 0 over 24 hours using logarithmic decay)
+
+## Version 1.0 (_August 2022_)
+
+* Latex Compatibility
+  * Used MathJax and Puppeteer to interpret Latex/HTML
+* Using HTML for better branding in problem of the day